// models/passenger.tsx
export default class Passenger {

    first_name?: string;
    middle_name?: string;
    last_name?: string;
    gender?: string;
    date_of_birth?: string;
    email?: string;
    cc?: string;
    mobile_number?: string;
    passport_number?: string;
    nationality?: string;
    issuing_nationality?: string;
    issuing_authority?: string;
    place_of_birth?: string;
    passport_issuance_date?: string;
    passport_expiry_date?: string;
}