// models/errors/passenger-form-error.tsx
export default class PassengerFormError {

    first_name?: string;
    middle_name?: string;
    last_name?: string;
    date_of_birth?: string;
    gender?: string;
    email?: string;
    cc?: string;
    mobile_number?: string;
    passport_number?: string;
    nationality?: string;
    issuing_authority?: string;
    issuing_nationality?: string;
    place_of_birth?: string;
    passport_issuance_date?: string;
    passport_expiry_date?: string;
    
}